feat(cities-skymine): agent conventions, cloud/bezier commands, MessageUtil migration (#743)
## Summary

- CitiesSkyMine の AI エージェント向け実装規約ドキュメントを3本新設（command-conventions /
implementation-conventions / best-practices）
- CloudCommand（非同期ノイズベース雲生成 + FAWE undo）と BezierCommand（3次ベジェ曲線道路 +
プレビュー）を追加
- RoadCurveCommand / HaussmannCommand の全メッセージを `§` コードから
MessageUtil（MiniMessage）に移行
- 3エージェント独立検証（blindspot mode）で発見された13項目（A〜M）をすべて修正:
  - Main.kt: cloudHandler フィールド昇格 + onDisable での cancelAll()
  - CloudCommand: cancelAll() メソッド追加
  - WindowCommand: getInt→getLong 一貫性修正
  - ColumnLayoutCommand / CitiesSkyMineCommand: ルーティング拡張
- ドキュメント: CsmEditSession overload 使い分け、limit key ドメイン固有性、async
判断基準、Session ライフサイクル分類、権限チェック所有権、bezier/preset ラベル書き換え等を追記

## Test plan

- [x] `nix develop --command gradle
:plugins:CitiesSkyMine:compileKotlin` ビルド成功を確認
- [ ] dev サーバーで `/csm cloud`, `/.cloud` の動作確認
- [ ] dev サーバーで `/csm bezier`, `/.bez` の動作確認
- [ ] `/rc build`, `/rc status`, `/rc set` の MessageUtil 移行後の表示確認
- [ ] `/hb build`, `/hb schem` の MessageUtil 移行後の表示確認
- [ ] サーバー reload 時に cloudHandler.cancelAll() が呼ばれることを確認


---------

+++ plugins/CitiesSkyMine/src/main/kotlin/icu/oyasai/citiesskymine/window/WindowCommand.kt
@@ -81,8 +81,8 @@ class WindowCommand(private val plugin: Main) : CommandExecutor, TabCompleter {
           emptyList()
         }
 
-    val operationBlocks = windowTargets.size + fillTargets.size
-    val maxBlocks = plugin.config.getInt("limits.max-blocks-window", 512)
+    val operationBlocks = (windowTargets.size + fillTargets.size).toLong()
+    val maxBlocks = plugin.config.getLong("limits.max-blocks-window", 512L)
     if (maxBlocks > 0 && operationBlocks > maxBlocks) {
       MessageUtil.error(sender, "生成ブロック数が上限 ($maxBlocks) を超えています: $operationBlocks")
       return true
